Moves submit to Faban functionality into FabanManagerService
