+++ benchflow-experiment-manager/application/src/main/java/cloud/benchflow/experimentmanager/tasks/running/ReExecuteTrialTask.java
@@ -1,6 +1,7 @@
 package cloud.benchflow.experimentmanager.tasks.running;
 
 import cloud.benchflow.experimentmanager.BenchFlowExperimentManagerApplication;
+import cloud.benchflow.experimentmanager.services.external.FabanManagerService;
 import cloud.benchflow.experimentmanager.services.external.MinioService;
 import cloud.benchflow.experimentmanager.services.internal.dao.TrialModelDAO;
 import cloud.benchflow.experimentmanager.tasks.running.execute.ExecuteTrial;
@@ -21,14 +22,14 @@ public class ReExecuteTrialTask implements Callable<TrialStatus> {
 
   private final String trialID;
   private TrialModelDAO trialModelDAO;
-  private MinioService minioService;
+  private FabanManagerService fabanManagerService;
   private FabanClient fabanClient;
 
   public ReExecuteTrialTask(String trialID) {
     this.trialID = trialID;
 
     this.trialModelDAO = BenchFlowExperimentManagerApplication.getTrialModelDAO();
-    this.minioService = BenchFlowExperimentManagerApplication.getMinioService();
+    this.fabanManagerService = BenchFlowExperimentManagerApplication.getFabanManagerService();
     this.fabanClient = BenchFlowExperimentManagerApplication.getFabanClient();
   }
 
@@ -40,6 +41,6 @@ public class ReExecuteTrialTask implements Callable<TrialStatus> {
     // get last executed trial
     trialModelDAO.incrementRetries(trialID);
 
-    return ExecuteTrial.executeTrial(trialID, trialModelDAO, minioService, fabanClient);
+    return ExecuteTrial.executeTrial(trialID, trialModelDAO, fabanManagerService, fabanClient);
   }
 }
