+++ benchflow-experiment-manager/application/src/main/java/cloud/benchflow/experimentmanager/tasks/running/execute/ExecuteTrial.java
@@ -4,6 +4,7 @@ import cloud.benchflow.experimentmanager.BenchFlowExperimentManagerApplication;
 import cloud.benchflow.experimentmanager.constants.BenchFlowConstants;
 import cloud.benchflow.experimentmanager.demo.DriversMakerCompatibleID;
 import cloud.benchflow.experimentmanager.exceptions.TrialIDDoesNotExistException;
+import cloud.benchflow.experimentmanager.services.external.FabanManagerService;
 import cloud.benchflow.experimentmanager.services.external.MinioService;
 import cloud.benchflow.experimentmanager.services.internal.dao.TrialModelDAO;
 import cloud.benchflow.faban.client.FabanClient;
@@ -20,7 +21,7 @@ import java.io.InputStream;
 import java.nio.charset.StandardCharsets;
 import java.nio.file.Paths;
 
-import static cloud.benchflow.experimentmanager.constants.BenchFlowConstants.getFabanTrialID;
+import static cloud.benchflow.experimentmanager.services.external.FabanManagerService.getFabanTrialID;
 import static cloud.benchflow.faban.client.responses.RunStatus.Code.*;
 
 /**
@@ -29,13 +30,11 @@ import static cloud.benchflow.faban.client.responses.RunStatus.Code.*;
 public class ExecuteTrial {
 
   public static TrialStatus executeTrial(String trialID, TrialModelDAO trialModelDAO,
-      MinioService minioService, FabanClient fabanClient) {
+      FabanManagerService fabanManagerService, FabanClient fabanClient) {
 
     try {
 
       String experimentID = BenchFlowConstants.getExperimentIDFromTrialID(trialID);
-
-      String fabanExperimentId = BenchFlowConstants.getFabanExperimentID(experimentID);
 
       DriversMakerCompatibleID driversMakerCompatibleID =
           new DriversMakerCompatibleID(experimentID);
@@ -44,53 +43,11 @@ public class ExecuteTrial {
 
       long experimentNumber = driversMakerCompatibleID.getExperimentNumber();
 
-      // A) submit to fabanClient
-      int submitRetries = BenchFlowExperimentManagerApplication.getSubmitRetries();
+      RunId runId = fabanManagerService.submitTrialToFaban(experimentID, trialID,
+          driversMakerExperimentID, experimentNumber);
 
-      int trialNumber = BenchFlowConstants.getTrialNumberFromTrialID(trialID);
-
-      java.nio.file.Path fabanConfigPath = Paths.get(BenchFlowConstants.TEMP_DIR)
-          .resolve(experimentID).resolve(String.valueOf(trialNumber))
-          .resolve(BenchFlowConstants.FABAN_CONFIGURATION_FILENAME);
-
-      InputStream configInputStream = minioService.getDriversMakerGeneratedFabanConfiguration(
-          driversMakerExperimentID, experimentNumber, trialNumber);
-      String config = IOUtils.toString(configInputStream, StandardCharsets.UTF_8);
-
-      FileUtils.writeStringToFile(fabanConfigPath.toFile(), config, StandardCharsets.UTF_8);
-
-      RunId runId = null;
-      while (runId == null) {
-        try {
-
-          // TODO - should this be a method (part of Faban Client?)
-          String fabanTrialId = getFabanTrialID(trialID);
-
-          runId = fabanClient.submit(fabanExperimentId, fabanTrialId, fabanConfigPath.toFile());
-
-        } catch (FabanClientException e) {
-
-          if (submitRetries > 0) {
-
-            // if there was an error submitting and we have not finished all retries
-            submitRetries--;
-
-          } else {
-            // TODO - handle me
-            throw e;
-          }
-        } catch (ConfigFileNotFoundException e) {
-          // TODO - handle me
-          e.printStackTrace();
-        }
-      }
-
-      // remove file that was sent to fabanClient
-      FileUtils.forceDelete(fabanConfigPath.toFile());
-
-      // TODO - when faban interaction changes and we don't have to do polling
-      trialModelDAO.setFabanTrialID(experimentID, trialNumber, runId.toString());
-      trialModelDAO.setTrialModelAsStarted(experimentID, trialNumber);
+      trialModelDAO.setFabanTrialID(trialID, runId.toString());
+      trialModelDAO.setTrialModelAsStarted(trialID);
 
       // B) wait/poll for trial to complete and store the trial result in the DB
       // TODO - is this the status we want to use? No it is a subset, should also include metrics computation status
